animations and better mixins

+++ common/src/main/java/net/mehvahdjukaar/sleep_tight/common/entities/BedEntity.java
@@ -440,7 +440,7 @@ public class BedEntity extends Entity implements IControllableVehicle, IExtraCli
             var d = player.getDeltaMovement();
             double vel = d.dot(MthUtils.V3itoV3(tile.getDirection().getClockWise().getNormal())) / d.length();
 
-            tile.addImpulse(-vel * 1.1f);
+            tile.addImpulse((float) (-vel * 1.1f));
         }
     }
